zsh: cleanup significant whitespace in _arguments
When an _arguments optspec uses a simple action (i.e. a command that is
invoked to generate matches, instead of `->action` or a list of matches
in parentheses), the whitespace preceding the action is significant:

>  word...
> word...
>
> This covers all forms other than those above.
> If the action starts with a space, the remaining list of words will be
> invoked unchanged.
>
> Otherwise it will be invoked with some extra strings placed after the
> first word; these are to be passed down as options to the compadd
> builtin. They ensure that the state specified by _arguments, in
> particular the descriptions of options and arguments, is correctly
> passed to the completion command.

Thus, remove the leading whitespace to ensure the description is passed
through to the function, except if aur-specific completion function
(_aur-*) is used as those do not handle arbitrary compadd options.

Link: https://zsh.sourceforge.io/Doc/Release/Completion-System.html