view: set orderFile on every call
This moves the `git orderFile` logic from `aur-sync` to `aur-view`.
Incidentally this was the only remaining `git` call in `aur-sync`.

Calling `git config orderFile` on every `aur-view` call, instead of
after a `git clone`, ensures `orderFile` always points to a valid path.
For example, if `HOME=/temphome/user` aur sync ...` is run for a series
of new(ly cloned) packages, subsequent calls `aur sync ...` using the
default `HOME=/home/user` value would still rely on the orderFile in
`/temphome/user`. This leads to errors if the path is deleted.

Closes #1167